refactor(objectstore): rename azure module to abs
