+++ crates/loonfs-cli/src/config.rs
@@ -1,7 +1,7 @@
 use crate::error::CliError;
 use http::Uri;
 use loonfs_api::NamespaceId;
-use loonfs_objectstore::azure::AzureAbsStoreConfig;
+use loonfs_objectstore::abs::AzureAbsStoreConfig;
 use loonfs_objectstore::gcs::GcsStoreConfig;
 use loonfs_objectstore::r2::R2StoreConfig;
 use loonfs_objectstore::s3::AwsS3StoreConfig;
